# group-7/project-2/backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "M1-QA 投研问答助手 API",
    "description": "对齐 spec/09 API 接口规格 — 支持 SSE 流式问答、会话管理、历史搜索、相似性检测、股票信息查询、指标对比、舆情监控、AI拉取配置",
    "version": "1.0.0"
  },
  "servers": [{ "url": "http://localhost:5000", "description": "本地开发" }],
...
  "paths": {
...
    "/api/v1/agent/compare": {
      "post": {
        "tags": ["指标对比"],
        "summary": "指标对比",
        "description": "对比指定股票在多份研报中的关键指标（评级、目标价、EPS、PE等），如无匹配研报则返回Demo数据",
        "operationId": "compareIndicators",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CompareRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "对比成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "traceId": { "type": "string" },
                    "stock_code": { "type": "string", "example": "600519" },
                    "stock_name": { "type": "string", "example": "贵州茅台" },
                    "indicators": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/CompareIndicator" }
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error400" }
        }
      }
    },
    "/api/v1/agent/compare/export": {
      "post": {
        "tags": ["指标对比"],
        "summary": "Excel导出",
        "description": "将指标对比结果导出为Excel文件（.xlsx），包含表头和数据行",
        "operationId": "exportCompare",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CompareRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Excel文件流",
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error400" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompareRequest": {
        "type": "object",
        "required": ["stock_code", "report_ids"],
        "properties": {
          "stock_code": { "type": "string", "description": "股票代码（6位）", "example": "600519" },
          "report_ids": { "type": "array", "items": { "type": "string" }, "description": "研报ID列表", "example": ["rpt_abc", "rpt_def"] }
        }
      },
      "CompareIndicator": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "description": "指标名称", "example": "评级" },
          "values": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "report_id": { "type": "string", "example": "demo_1" },
                "report_title": { "type": "string", "example": "华泰证券研报" },
                "value": { "type": "string", "example": "买入" },
                "source_text": { "type": "string", "example": "维持买入评级，看好公司长期发展前景" }
              }
            }
          }
        }
      },
...
    },
    "responses": {
      "Error400": {
        "description": "请求错误",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "string", "example": "EMPTY_QUERY" },
                    "message": { "type": "string", "example": "请输入问题" },
                    "details": { "type": "object" },
                    "traceId": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}